[util] Extract LogDisplay component into a new utility
