+++ src/services/memory.tsx
@@ -20,6 +20,7 @@ import { DiscoveryClient } from 'services/client/discover';
 import { fromFixed, MemoryAllocator, Worker } from 'services/allocator';
 
 import { useNsUpdate, useTheme } from 'util/hooks';
+import { LogDisplay } from 'ui/LogDisplay';
 import { BaseServer, Handlers } from 'util/protocol';
 import { HUD_HEIGHT, HUD_WIDTH, STATUS_WINDOW_WIDTH } from 'util/ui';
 
@@ -384,55 +385,6 @@ function MemoryDisplay({ ns, getWorkers }: MemoryDisplayProps) {
                     ))}
                 </tbody>
             </table>
-        </div>
-    );
-}
-
-interface LogDisplayProps {
-    ns: NS;
-    lines: string[];
-}
-
-/**
- * Show recent log lines in a styled list.
- *
- * @param lines - Log messages to display.
- * @param theme - The UI theme.
- */
-function LogDisplay({ ns, lines: extLines }: LogDisplayProps) {
-    const [lines, setLines] = React.useState(extLines);
-    const theme = useTheme(ns, 1000);
-
-    React.useEffect(() => {
-        const id = globalThis.setInterval(() => {
-            setLines(extLines);
-        }, 1000);
-        return () => {
-            globalThis.clearInterval(id);
-        };
-    }, [ns, extLines]);
-
-    const rowStyle = (idx: number) =>
-        idx % 2 === 1 ? { backgroundColor: theme.well } : {};
-
-    const lineColor = (line: string): string | undefined => {
-        if (line.startsWith('ERROR:')) return theme.error;
-        if (line.startsWith('SUCCESS:')) return theme.success;
-        if (line.startsWith('WARN:')) return theme.warning;
-        if (line.startsWith('INFO:')) return theme.info;
-        return theme.success;
-    };
-
-    return (
-        <div style={{ fontFamily: 'monospace' }}>
-            {lines.map((line, idx) => {
-                const style = { ...rowStyle(idx), color: lineColor(line) };
-                return (
-                    <div key={idx} style={style}>
-                        {line}
-                    </div>
-                );
-            })}
         </div>
     );
 }
